Update better_asteroids
Rather crappy. I tried to get the pellet to work, but it seems like objects called within functions stay within functions and do not apply globally. Had to use an if statement instead of the keyPressed function. Ugh... I thought that I almost got a pellet to shoot out. I'll probably need to create a "dummy" pellet that's simply not drawn in order to get this function to work. Probably an if statement with a Boolean value should work. Also, objects need to be updated individually. A function in the class apparently doesn't update all objects under the class but only apply to the single object called. Oh, by the way, the current code give a  "NullPointerException" one the pellet.p_update() line and says that the pellet object is a local variable + is not used :(